Cleaned up tests and removed unused imports
