+++ src/index.ts
@@ -1,8 +1,5 @@
-// import * as DotEnv from 'dotenv';
 import Server from './server';
-import * as Hapi from 'hapi';
 
-// DotEnv.config();
 (async () => {
     await Server.start();
 })();
